Allow syntax lines to set an action number independent of the action routine: <SYNTAX FOO OBJECT = V-FOO>  ;"PRSA = V?FOO" <SYNTAX FOO OBJECT WITH OBJECT = V-FOO <> FOO-WITH>  ;"PRSA = V?FOO-WITH"
Preactions are linked to the action number, not the action routine, so V?FOO and
V?FOO-WITH can have different preactions.

Also fixed the order in which syntax lines are processed. They're now (again)
read in definition order so that the earliest preaction will stick, but the
resulting syntax data is still emitted in reverse order so that the game
will use later lines first.