Fixed AGAIN for orphaned commands.
BACKLEXBUF and BACKREADBUF are absorbed into a PARSER-RESULT structure that
also stores the matched objects, verb number, etc., and AGAIN repeats the parsed
command instead of reparsing the text. We do this because it's possible to use
orphaning to perform actions that can't be expressed in a single command (if
there are objects that can't be named unambiguously with a single adjective/noun
pair).

Unfortunately(?), this means ANY won't re-randomize after typing AGAIN.