Fixed bug where ZILF didn't escape/unescape exotic atom names consistently when using them as ZAP symbols, sometimes producing assembly code with missing or misspelled names.
For example, ZILF 0.8 generates these table names for the fake verb ",tell"...

W?$002cTELL         (an encoding of "W?,TELL")
ST?$005c$002cTELL   (an encoding of "ST?\,TELL")
ACT?$005c$002cTELL

As long as the spelling of each specific table name was consistent, this was
only a cosmetic problem. But in rare circumstances, two parts of the code would
have different ideas of how to escape the same name, leading to assembly errors.